+++ pkgs/hooks_runner/test/locking/locking_test.dart
@@ -7,6 +7,7 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:math';
 
+import 'package:native_test_helpers/native_test_helpers.dart';
 import 'package:test/test.dart';
 
 import '../helpers.dart';