+++ pkgs/native_toolchain_c/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.16.7
+
+* Support Module Definitions for linking on Windows.
+
+## 0.16.6
+
+* Support linking for Windows.
+
 ## 0.16.5
 
 * Support linking for iOS.